web-server/requirements.txt
astroid==2.4.2
branca==0.4.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
earthengine-api==0.1.238
Flask==1.1.2
Flask-Cors==3.0.9
folium==0.11.0
future==0.18.2
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
httplib2==0.18.1
httplib2shim==0.0.3
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
